Move PLP/row-writer streaming machinery out of api into fetch_engine
The api module is reserved for ODBC API entry points. plp_stream and
row_writer are internal column-wise fetch machinery, so move them into a
new fetch_engine module and update all references.
