+++ mssql-odbc/src/api/mod.rs
@@ -18,9 +18,7 @@ mod get_type_info;
 mod more_results;
 mod num_result_cols;
 pub(crate) mod odbc_types;
-pub(crate) mod plp_stream;
 mod prepare;
-pub(crate) mod row_writer;
 pub(crate) mod set_connect_attr;
 pub(crate) mod set_env_attr;
 pub(crate) mod set_stmt_attr;
